Rebase: Merge BOLT codebase in monorepo
Summary:
This commit is the first step in rebasing all of BOLT
history in the LLVM monorepo. It also solves trivial build issues
by updating BOLT codebase to use current LLVM. There is still work
left in rebasing some BOLT features and in making sure everything
is working as intended.

History has been rewritten to put BOLT in the /bolt folder, as
opposed to /tools/llvm-bolt.

(cherry picked from FBD33289252)

+++ bolt/src/DWARFRewriter.h
@@ -53,7 +53,7 @@ class DWARFRewriter {
   std::vector<std::unique_ptr<DebugLocWriter>> LocListWritersByCU;
 
   struct LocListDebugInfoPatchType {
-    uint32_t DebugInfoOffset;
+    uint64_t DebugInfoOffset;
     size_t CUIndex;
     uint64_t CUWriterOffset;
   };
